# backend/salonbw-backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products/{id}": {
      "get": {
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get product by id",
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product",
        "tags": [
          "products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product removed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove product",
        "tags": [
          "products"
        ]
      }
    },
...
    "/product-categories": {
      "get": {
        "operationId": "ProductCategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCategory"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List product categories",
        "tags": [
          "product-categories"
        ]
      },
      "post": {
        "operationId": "ProductCategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategory"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product category",
        "tags": [
          "product-categories"
        ]
      }
    },
    "/product-categories/tree": {
      "get": {
        "operationId": "ProductCategoriesController_findTree",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductCategory"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get product categories as tree",
        "tags": [
          "product-categories"
        ]
      }
    },
    "/product-categories/{id}": {
      "get": {
        "operationId": "ProductCategoriesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategory"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get product category by id",
        "tags": [
          "product-categories"
        ]
      },
      "patch": {
        "operationId": "ProductCategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategory"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update product category",
        "tags": [
          "product-categories"
        ]
      },
      "delete": {
        "operationId": "ProductCategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product category removed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete product category",
        "tags": [
          "product-categories"
        ]
      }
    },
...
    "/usage": {
      "post": {
        "operationId": "UsageController_createUsage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usage recorded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create product usage (material consumption)",
        "tags": [
          "retail"
        ]
      },
      "get": {
        "operationId": "UsageController_findUsage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Usage records"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List product usage records",
        "tags": [
          "retail"
        ]
      }
    },
    "/usage/{id}": {
      "get": {
        "operationId": "UsageController_findOneUsage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage details"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get usage details",
        "tags": [
          "retail"
        ]
      }
    },
...
  },
  "info": {
    "title": "SalonBW API",
    "description": "The SalonBW API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Product": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "unitPrice": {
            "type": "number"
          },
          "vatRate": {
            "type": "number"
          },
          "stock": {
            "type": "number"
          },
          "purchasePrice": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "sku": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "productType": {
            "type": "string",
            "enum": [
              "product",
              "supply",
              "universal"
            ]
          },
          "minQuantity": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "packageSize": {
            "type": "number"
          },
          "packageUnit": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "defaultSupplierId": {
            "type": "number"
          },
          "categoryId": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          },
          "trackStock": {
            "type": "boolean"
          }
        }
      },
      "ProductCategory": {
        "type": "object",
        "properties": {}
      },
      "CreateProductCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "number"
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateProductCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "number"
          },
          "sortOrder": {
            "type": "number"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "CreateUsageItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number"
          },
          "quantity": {
            "type": "number",
            "minimum": 1
          },
          "unit": {
            "type": "string",
            "default": "op."
          }
        },
        "required": [
          "productId",
          "quantity"
        ]
      },
      "CreateUsageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateUsageItemDto"
            }
          },
          "employeeId": {
            "type": "number"
          },
          "appointmentId": {
            "type": "number"
          },
          "clientName": {
            "type": "string"
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}